[add] planner initializer & test code
