+++ src/main/kotlin/org/pasudo123/commitsearcher/config/CacheConfiguration.kt
@@ -1,0 +1,9 @@
+package org.pasudo123.commitsearcher.config
+
+import org.springframework.cache.annotation.EnableCaching
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+@EnableCaching
+class CacheConfiguration {
+}
