+++ src/test/kotlin/org/pasudo123/commitsearcher/til/controller/TilControllerTest.kt
@@ -81,6 +81,7 @@ internal class TilControllerTest {
         }.andExpect {
             status { isOk }
             content { contentType(MediaType.APPLICATION_JSON) }
+            content { json("{\"myContribution\":{}, \"html\":{}}") }
         }.andDo {
             print()
         }.andReturn()
